+++ app/build.gradle.kts
@@ -3,9 +3,6 @@ import com.android.*
 import java.util.Properties
 
 plugins {
-    id("com.google.dagger.hilt.android")
-    kotlin("kapt")
-
     id("com.android.application")
     id("org.jetbrains.kotlin.android")
     id("com.google.dagger.hilt.android")
