+++ app/src/main/java/com/example/teacherforboss/presentation/ui/auth/signup/fragment/AgreementMoreFragment.kt
@@ -1,4 +1,4 @@
-package com.example.teacherforboss.signup.fragment
+package com.example.teacherforboss.presentation.ui.auth.signup.fragment
 
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
@@ -10,19 +10,18 @@ import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.viewModels
 import com.example.teacherforboss.R
 import com.example.teacherforboss.databinding.FragmentAgreementMoreBinding
-import com.example.teacherforboss.databinding.FragmentPersonalinfoMoreBinding
-import com.example.teacherforboss.signup.SignupViewModel
+import com.example.teacherforboss.presentation.ui.auth.signup.SignupViewModel
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 
-class PersonalInfoMoreFragment : BottomSheetDialogFragment() {
-    private lateinit var binding: FragmentPersonalinfoMoreBinding
+class AgreementMoreFragment : BottomSheetDialogFragment() {
+    private lateinit var binding: FragmentAgreementMoreBinding
     private val viewModel: SignupViewModel by viewModels()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding= DataBindingUtil.inflate(inflater, R.layout.fragment_personalinfo_more,container,false)
+        binding= DataBindingUtil.inflate(inflater, R.layout.fragment_agreement_more,container,false)
 
         binding.signupViewModel=viewModel
         binding.lifecycleOwner=this
