+++ app/src/main/java/com/example/teacherforboss/presentation/ui/auth/signup/fragment/GenderBirthFragment.kt
@@ -5,17 +5,15 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.CheckBox
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.viewModels
 import com.example.teacherforboss.R
 import com.example.teacherforboss.R.style.AppBottomSheetDialogTheme
-import com.example.teacherforboss.databinding.FragmentEmailBinding
 import com.example.teacherforboss.databinding.FragmentGenderBirthBinding
-import com.example.teacherforboss.signup.SignupActivity
-import com.example.teacherforboss.signup.SignupViewModel
-
+import com.example.teacherforboss.presentation.ui.auth.signup.SignupActivity
+import com.example.teacherforboss.presentation.ui.auth.signup.SignupViewModel
+import com.example.teacherforboss.presentation.ui.auth.signup.fragment.AgreementFragment
 
 class GenderBirthFragment : Fragment() {
     private lateinit var binding: FragmentGenderBirthBinding
@@ -37,7 +35,7 @@ class GenderBirthFragment : Fragment() {
         val activity=activity as SignupActivity
 
         binding.nextBtn.setOnClickListener {
-            val bottomSheetDialog=AgreementFragment()
+            val bottomSheetDialog= AgreementFragment()
             bottomSheetDialog.setStyle(DialogFragment.STYLE_NORMAL, AppBottomSheetDialogTheme)
             bottomSheetDialog.show(activity.supportFragmentManager,"agreement")
         }
