+++ app/src/main/java/com/example/teacherforboss/presentation/ui/auth/signup/fragment/LocationServiceMoreFragment.kt
@@ -1,4 +1,4 @@
-package com.example.teacherforboss.signup.fragment
+package com.example.teacherforboss.presentation.ui.auth.signup.fragment
 
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
@@ -9,19 +9,20 @@ import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.viewModels
 import com.example.teacherforboss.R
-import com.example.teacherforboss.databinding.FragmentAgreementMoreBinding
-import com.example.teacherforboss.signup.SignupViewModel
+import com.example.teacherforboss.databinding.FragmentLocationserviceMoreBinding
+import com.example.teacherforboss.databinding.FragmentPersonalinfoMoreBinding
+import com.example.teacherforboss.presentation.ui.auth.signup.SignupViewModel
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 
-class AgreementMoreFragment : BottomSheetDialogFragment() {
-    private lateinit var binding: FragmentAgreementMoreBinding
+class LocationServiceMoreFragment : BottomSheetDialogFragment() {
+    private lateinit var binding: FragmentLocationserviceMoreBinding
     private val viewModel: SignupViewModel by viewModels()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding= DataBindingUtil.inflate(inflater, R.layout.fragment_agreement_more,container,false)
+        binding= DataBindingUtil.inflate(inflater, R.layout.fragment_locationservice_more,container,false)
 
         binding.signupViewModel=viewModel
         binding.lifecycleOwner=this
